+++ site/.gitignore
@@ -1,4 +1,21 @@
-/node_modules/
-/public/build/
+node_modules
 
+# Output
+.output
+.vercel
+/.svelte-kit
+/build
+
+# OS
 .DS_Store
+Thumbs.db
+
+# Env
+.env
+.env.*
+!.env.example
+!.env.test
+
+# Vite
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
