--- site/src/global.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="svelte" />
